requirements.txt
# ─── requirements.txt ────────────────────────────────────────────────────────
anthropic>=0.25.0
openai>=1.12.0
google-generativeai>=0.5.0
datasets>=2.19.0
transformers>=4.40.0
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
tree-sitter==0.20.4